+++ tests/test_task3/test_rpq.json
@@ -1,0 +1,98 @@
+{
+  "rpq": [
+    {
+      "graph": "two_cycle_graph",
+      "query": "a*|b*",
+      "starts": [0, 1, 2, 3, 4, 5],
+      "finishes": [0, 1, 2, 3, 4, 5],
+      "expected": [
+                [4, 0],
+                [3, 1],
+                [5, 4],
+                [0, 2],
+                [0, 5],
+                [2, 2],
+                [1, 0],
+                [1, 3],
+                [3, 0],
+                [4, 5],
+                [3, 3],
+                [5, 0],
+                [0, 1],
+                [1, 2],
+                [0, 4],
+                [2, 1],
+                [3, 2],
+                [4, 4],
+                [5, 5],
+                [0, 0],
+                [1, 1],
+                [0, 3],
+                [2, 0],
+                [2, 3]
+            ]
+    },
+    {
+      "graph": "two_cycle_graph",
+      "query": "b*",
+      "starts": [0],
+      "finishes": [5, 4],
+      "expected": [
+        [0, 5], [0, 4]
+      ]
+    },
+    {
+      "graph": "two_cycle_graph",
+      "query": "a*|b*",
+      "starts": [0],
+      "finishes": [1, 2, 3, 4],
+      "expected": [
+        [0, 1], [0, 2], [0, 3], [0, 4]
+      ]
+    },
+    {
+      "graph": "two_cycle_graph",
+      "query": "a*|b*",
+      "starts": [4],
+      "finishes": [4, 5],
+      "expected": [
+        [4, 4], [4, 5]
+      ]
+    },
+    {
+      "graph": "two_cycle_graph",
+      "query": "b",
+      "starts": [0],
+      "finishes": [0, 1, 2, 3],
+      "expected": []
+    },
+    {
+      "graph": "two_cycle_graph",
+      "query": "b*",
+      "starts": [0],
+      "finishes": [5, 4],
+      "expected": [[0, 5], [0, 4]]
+    },
+    {
+      "graph": "empty_graph",
+      "query": "",
+      "starts": [],
+      "finishes": [],
+      "expected": []
+    },
+    {
+      "graph": "empty_graph",
+      "query": "a*|b*",
+      "starts": [],
+      "finishes": [],
+      "expected": []
+    },
+    {
+      "graph": "linear_graph",
+      "query": "a b b",
+      "starts": [0, 1, 2, 3],
+      "finishes": [0, 1, 2, 3],
+      "expected": [[0, 3]]
+    }
+  ]
+}
